requirements.txt
pyserial
Flask
requests
matplotlib
openpyxl